Rewritten PDF module.
The pdf tool was using the pdf2swf for viewing, and provided no creation
of pdf files. The new version does not provide a conversion for viewing
as the major browsers already have good support for that. Creating a pdf
from a number of images is now possible (/asset/id/pdf/create), as well
generating a still from a pdf.